Move all MusicLibraryMenu Changes to their own class.
